+++ test/cases/deprecate-option-extractCss/webpack.config.js
@@ -1,0 +1,37 @@
+const path = require('path');
+const PugPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+  devtool: false,
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+  },
+
+  entry: {
+    index: './src/views/index.pug',
+  },
+
+  plugins: [
+    new PugPlugin({
+      // test deprecation for the extractCss option
+      extractCss: {
+        filename: '[name].[contenthash:8].css',
+      },
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.pug$/,
+        loader: PugPlugin.loader,
+      },
+      {
+        test: /\.(css|sass|scss)$/,
+        use: ['css-loader'],
+      },
+    ],
+  },
+};