+++ test/cases/entry-styles-bypass/webpack.config.js
@@ -23,7 +23,7 @@ module.exports = {
 
   entry: {
     index: './src/views/index.pug',
-    styles: './src/assets/styles/main.scss',
+    style: './src/assets/styles/main.scss',
     about: './src/assets/styles/about.css',
   },