+++ test/cases/inline-asset-bypass-data-url/webpack.config.js
@@ -1,0 +1,34 @@
+const path = require('path');
+const PugPlugin = require('../../../');
+
+module.exports = {
+  mode: 'development',
+  devtool: 'source-map',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+    publicPath: '/',
+  },
+
+  entry: {
+    index: './src/views/index.pug',
+  },
+
+  plugins: [new PugPlugin()],
+
+  module: {
+    rules: [
+      {
+        test: /\.(pug)$/,
+        loader: PugPlugin.loader,
+        options: {
+          method: 'render',
+        },
+      },
+      {
+        test: /\.(css|sass|scss)$/,
+        use: ['css-loader', 'sass-loader'],
+      },
+    ],
+  },
+};