+++ test/cases/inline-script-query/webpack.config.js
@@ -1,0 +1,37 @@
+const path = require('path');
+const PugPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+  },
+
+  resolve: {
+    alias: {
+      Scripts: path.join(__dirname, 'src/assets/scripts/'),
+    },
+  },
+
+  entry: {
+    index: './src/views/index.pug',
+  },
+
+  plugins: [
+    new PugPlugin({
+      js: {
+        filename: 'assets/js/[name].[contenthash:8].js',
+      },
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.pug$/,
+        loader: PugPlugin.loader,
+      },
+    ],
+  },
+};