+++ test/cases/resolve-url-in-css-render/webpack.config.js
@@ -1,0 +1,52 @@
+const path = require('path');
+const PugPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+  devtool: false,
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+    publicPath: '/',
+  },
+
+  entry: {
+    index: './src/index.pug',
+  },
+
+  plugins: [
+    new PugPlugin({
+      js: {
+        filename: 'assets/js/[name].[contenthash:8].js',
+      },
+      css: {
+        filename: 'assets/css/[name].[contenthash:8].css',
+      },
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.pug$/,
+        loader: PugPlugin.loader,
+        options: {
+          method: 'render',
+        },
+      },
+
+      {
+        test: /\.(css)$/,
+        loader: 'css-loader',
+      },
+
+      {
+        test: /\.(png|jpe?g|ico)/,
+        type: 'asset/resource',
+        generator: {
+          filename: 'assets/img/[name].[hash:8][ext]',
+        },
+      },
+    ],
+  },
+};